Update datacard
The main issue is the process name mixmatch between json, conf and
the datacard. I created a nested map to keep track of them. We can
now have the systematics of photon in the photon CR.

For the QCDCR, I also force negative yeild to 0 for all the processes,
including rare.  Rare is now correlated with diboson